backend/requirements.txt
flask==3.1.0
sqlalchemy==2.0.38
psycopg2-binary==2.9.10
speechrecognition==3.14.1
spacy==3.7.6
cosmpy==0.9.2
pyyaml==6.0.2
jinja2==3.1.4
ecdsa==0.19.0
python-dotenv==1.0.1
redis==6.2.0
flask_limiter==3.12
